Fix writing to servos
This fixes an issue where the pulse width for a servo would be too
large an cause it to stall out.  This was because the pulse width
was not being caused by a percentage of the total number of degrees,
but instead using the total number of degrees.

The number of degrees as 180 has been determined based on testing
where it was discovered that 180 degrees is typically the maximum
angle for a servo.

Originally noted in https://github.com/hybridgroup/cylon-intel-iot/commit/122ff58a2a3a4052cebef901aa1eaeb849eecdad#commitcomment-10680946.

Signed-off-by: Kevin Brown <kevin@kevinbrown.in>